Refactor HasPrivacy onto Property
