+++ betty/plugins/entity/source.py
@@ -20,9 +20,11 @@ from betty.linked_data import JsonLdObject, dump_context
 from betty.locale.localizable.gettext import _, ngettext
 from betty.locale.localizable.linked_data import dump_linked_data
 from betty.locale.localizable.static.schema import StaticTranslationsSchema
-from betty.privacy import HasPrivacy, Privacy, is_public, merge_privacies
+from betty.privacy import Privacy
+from betty.privacy.resolve import is_public, merge_privacies
 from betty.properties.date import HasAnyDate
 from betty.properties.localizable import LocalizableProperty
+from betty.properties.privacy import HasPrivacy
 from betty.property import Optional
 
 if TYPE_CHECKING:
@@ -106,7 +108,7 @@ class Source(HasAnyDate, HasFileReferences, HasNotes, HasLinks, HasPrivacy, Enti
         date: AnyDate | None = None,
         file_references: ToManyAssociates[FileReference] = (),
         links: ToManyAssociates[Link] = (),
-        privacy: Privacy | None = None,
+        privacy: Privacy = Privacy.UNDETERMINED,
     ):
         super().__init__(
             id,
