+++ betty/privacy/privatizer.py
@@ -21,11 +21,12 @@ from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
 from betty.plugins.event_type.death import Death
 from betty.plugins.role.subject import Subject
-from betty.privacy import HasPrivacy, Privacy
+from betty.privacy import Privacy
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, MutableSequence
 
+    from betty.properties.privacy import HasPrivacy
     from betty.typing import Intersection
     from betty.user import User
