+++ betty/test_utils/privacy.py
@@ -2,10 +2,10 @@
 Test utilities for :py:mod:`betty.privacy`.
 """
 
-from betty.privacy import HasPrivacy
+from betty.properties.privacy import HasPrivacy
 
 
 class DummyHasPrivacy(HasPrivacy):
     """
-    A dummy :py:class:`betty.privacy.HasPrivacy` implementation.
+    A dummy :py:class:`betty.properties.privacy.HasPrivacy` implementation.
     """
