+++ betty/tests/plugins/entity/test_citation.py
@@ -97,7 +97,7 @@ class TestCitation(EntityTestBase):
             "@id": "https://example.com/citation/the_citation/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_citation",
-            "private": False,
+            "privacy": False,
             "facts": [],
             "links": [],
             "fileReferences": [],
@@ -127,7 +127,7 @@ class TestCitation(EntityTestBase):
             "@id": "https://example.com/citation/the_citation/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_citation",
-            "private": False,
+            "privacy": False,
             "location": {DEFAULT_LOCALE_TAG: "My First Location"},
             "source": "/source/the_source/index.json",
             "facts": ["/event/the_event/index.json"],
@@ -158,7 +158,7 @@ class TestCitation(EntityTestBase):
             "@id": "https://example.com/citation/the_citation/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_citation",
-            "private": True,
+            "privacy": True,
             "source": "/source/the_source/index.json",
             "facts": ["/event/the_event/index.json"],
             "links": [],
