+++ betty/tests/plugins/entity/test_event.py
@@ -146,7 +146,7 @@ class TestEvent(EntityTestBase):
             "@id": "https://example.com/event/the_event/index.json",
             "@type": "https://schema.org/Event",
             "id": "the_event",
-            "private": False,
+            "privacy": False,
             "type": "birth",
             "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
             "eventStatus": "https://schema.org/EventScheduled",
@@ -193,7 +193,7 @@ class TestEvent(EntityTestBase):
             "@id": "https://example.com/event/the_event/index.json",
             "@type": "https://schema.org/Event",
             "id": "the_event",
-            "private": False,
+            "privacy": False,
             "type": "birth",
             "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
             "eventStatus": "https://schema.org/EventScheduled",
@@ -203,7 +203,7 @@ class TestEvent(EntityTestBase):
                     "role": "subject",
                     "person": "/person/the_person/index.json",
                     "event": "/event/the_event/index.json",
-                    "private": False,
+                    "privacy": False,
                 },
             ],
             "citations": [
@@ -273,7 +273,7 @@ class TestEvent(EntityTestBase):
             "@id": "https://example.com/event/the_event/index.json",
             "@type": "https://schema.org/Event",
             "id": "the_event",
-            "private": True,
+            "privacy": True,
             "type": "birth",
             "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
             "eventStatus": "https://schema.org/EventScheduled",
@@ -282,7 +282,7 @@ class TestEvent(EntityTestBase):
                     "id": presence.id,
                     "person": "/person/the_person/index.json",
                     "event": "/event/the_event/index.json",
-                    "private": True,
+                    "privacy": True,
                 },
             ],
             "citations": [
