+++ betty/tests/plugins/entity/test_source.py
@@ -138,7 +138,7 @@ class TestSource(EntityTestBase):
             "@id": "https://example.com/source/the_source/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_source",
-            "private": False,
+            "privacy": False,
             "name": {DEFAULT_LOCALE_TAG: "The Source"},
             "fileReferences": [],
             "contains": [],
@@ -184,7 +184,7 @@ class TestSource(EntityTestBase):
             "@id": "https://example.com/source/the_source/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_source",
-            "private": False,
+            "privacy": False,
             "name": {DEFAULT_LOCALE_TAG: "The Source"},
             "author": {DEFAULT_LOCALE_TAG: "The Author"},
             "publisher": {DEFAULT_LOCALE_TAG: "The Publisher"},
@@ -217,7 +217,7 @@ class TestSource(EntityTestBase):
                         DEFAULT_LOCALE_TAG: "The Source Online",
                     },
                     "owner": "/source/the_source/index.json",
-                    "private": False,
+                    "privacy": False,
                 },
             ],
         }
@@ -259,7 +259,7 @@ class TestSource(EntityTestBase):
             "@id": "https://example.com/source/the_source/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_source",
-            "private": True,
+            "privacy": True,
             "fileReferences": [],
             "contains": [
                 "/source/the_contained_source/index.json",
@@ -304,7 +304,7 @@ class TestSource(EntityTestBase):
             "@id": "https://example.com/source/the_source/index.json",
             "@type": "https://schema.org/Thing",
             "id": "the_source",
-            "private": False,
+            "privacy": False,
             "fileReferences": [],
             "contains": [
                 "/source/the_contained_source/index.json",
