+++ data/asset/builtin/locale/betty.pot
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Betty VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2026-05-04 23:28+0100\n"
+"POT-Creation-Date: 2026-05-05 13:39+0100\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -1351,6 +1351,9 @@ msgid "Previous page"
 msgstr ""
 
 msgid "Primary color"
+msgstr ""
+
+msgid "Privacy"
 msgstr ""
 
 #, python-brace-format
