+++ data/asset/builtin/locale/nl-NL/betty.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2026-05-04 23:28+0100\n"
+"POT-Creation-Date: 2026-05-05 13:39+0100\n"
 "PO-Revision-Date: 2026-02-17 19:37+0000\n"
 "Last-Translator: Bart Feenstra <bart@bartfeenstra.com>\n"
 "Language: nl_NL\n"
@@ -1424,6 +1424,9 @@ msgstr "Vorige pagina"
 
 msgid "Primary color"
 msgstr "Primaire kleur"
+
+msgid "Privacy"
+msgstr ""
 
 #, python-brace-format
 msgid ""
